% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-australia.R
\name{get_debates}
\alias{get_debates}
\title{Retrieve summary data Hansard debates for a given day from the Open Australia API}
\usage{
get_debates(
  date = Sys.Date(),
  type = c("representatives", "senate"),
  api_key = Sys.getenv("OPEN_AUSTRALIA_KEY")
)
}
\arguments{
\item{date}{Retrieve debates for the given day. Defaults to the current day.
If given as a character, this function will attempt to convert it into a
date.}

\item{type}{Either "representatives" or "senate". Defaults to
"representatives".}

\item{api_key}{The Open Australia API key, available at
\url{https://www.openaustralia.org.au/api/}. Defaults to the value stored
in the "OPEN_AUSTRALIA_KEY" environment variable.}
}
\value{

}
\description{
Retrieve summary data Hansard debates for a given day from the Open Australia API
}
