% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-hansard.R
\name{hansard_link_pattern}
\alias{hansard_link_pattern}
\title{Define a regular expression pattern for matching links to Hansard pages}
\usage{
hansard_link_pattern()
}
\value{
Character
}
\description{
Define a regular expression pattern for matching links to Hansard pages
}
